+++ client/slack-spring-api-client/src/main/kotlin/io/olaph/slack/client/spring/group/users/DefaultUserListMethod.kt
@@ -1,6 +1,5 @@
 package io.olaph.slack.client.spring.group.users
 
-
 import io.olaph.slack.client.group.ApiCallResult
 import io.olaph.slack.client.group.users.UserListMethod
 import io.olaph.slack.client.spring.group.RestTemplateFactory
@@ -9,7 +8,6 @@ import io.olaph.slack.dto.jackson.group.users.ErrorUserListResponse
 import io.olaph.slack.dto.jackson.group.users.SuccessfulUserListResponse
 import io.olaph.slack.dto.jackson.group.users.UserListResponse
 import org.springframework.web.client.RestTemplate
-
 
 @Suppress("UNCHECKED_CAST")
 class DefaultUserListMethod(private val authToken: String, private val restTemplate: RestTemplate = RestTemplateFactory.slackTemplate()) : UserListMethod() {
@@ -24,6 +22,10 @@ class DefaultUserListMethod(private val authToken: String, private val restTempl
             is SuccessfulUserListResponse -> {
                 val responseEntity = response.body as SuccessfulUserListResponse
                 this.onSuccess?.invoke(responseEntity)
+                while (responseEntity.responseMetadata.nextCursor != "") {
+                    this.params.copy(cursor = responseEntity.responseMetadata.nextCursor)
+                    this.onSuccess?.invoke(responseEntity)
+                }
                 ApiCallResult(success = responseEntity)
             }
             is ErrorUserListResponse -> {
