+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/common/types/IdentityExtension.kt
@@ -1,0 +1,7 @@
+package io.olaph.slack.dto.jackson.common.types
+
+fun Identity.User.Companion.sample() = Identity.User("", "", "", "", "", "", "", "")
+
+fun Identity.Team.Companion.sample() = Identity.Team("", "")
+
+fun Identity.Companion.sample() = Identity(Identity.User.sample(), Identity.Team.sample())
