+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/common/types/MemberExtension.kt
@@ -1,0 +1,4 @@
+package io.olaph.slack.dto.jackson.common.types
+
+fun Member.Companion.sample() = Member("", "", "", false, "", "", "", "", 0, UserProfile.sample(),
+        false, false, false, false, false, false, 0, false, false, "")
