+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/common/types/UserProfileExtension.kt
@@ -1,0 +1,6 @@
+package io.olaph.slack.dto.jackson.common.types
+
+fun UserProfile.Companion.sample() = UserProfile("", "", "", "", "",
+        "", "", mapOf(), "", "",0, "",
+        false, "", "", "", "", "", "", "",
+        "", "", "", "", "", "")
