+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/group/usergroups/users/UsergroupsUsersListExtension.kt
@@ -1,9 +1,6 @@
 package io.olaph.slack.dto.jackson.group.usergroups.users
 
-import io.olaph.slack.dto.jackson.group.users.User
-import io.olaph.slack.dto.jackson.group.users.sample
-
-fun SuccessfulUsergroupsUsersListResponse.Companion.sample() = SuccessfulUsergroupsUsersListResponse(true, listOf(User.sample()))
+fun SuccessfulUsergroupsUsersListResponse.Companion.sample() = SuccessfulUsergroupsUsersListResponse(true, listOf())
 
 fun ErrorUsergroupsUsersListResponse.Companion.sample() = ErrorUsergroupsUsersListResponse(false, "")
