+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/group/users/UsersGetProfile.kt
@@ -1,17 +1,14 @@
 package io.olaph.slack.dto.jackson.group.users
 
+import io.olaph.slack.dto.jackson.common.types.UserProfile
+import io.olaph.slack.dto.jackson.common.types.sample
+
 fun SuccessfulUsersGetProfileResponse.Companion.sample(): SuccessfulUsersGetProfileResponse {
-    return SuccessfulUsersGetProfileResponse(true, SuccessfulUsersGetProfileResponse.Profile.sample())
+    return SuccessfulUsersGetProfileResponse(true, UserProfile.sample())
 }
 
 fun ErrorUsersGetProfileResponse.Companion.sample(): ErrorUsersGetProfileResponse {
     return ErrorUsersGetProfileResponse(true, "")
-}
-
-fun SuccessfulUsersGetProfileResponse.Profile.Companion.sample(): SuccessfulUsersGetProfileResponse.Profile {
-    return SuccessfulUsersGetProfileResponse.Profile("", "", "", "", "", "", "",
-            "", "", "", 0, "", "", "", "", "", "", "",
-            "", "", "")
 }
 
 fun UsersGetProfileRequest.Companion.sample(): UsersGetProfileRequest {
