+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/group/users/UsersLookupByEmailExtension.kt
@@ -1,6 +1,9 @@
 package io.olaph.slack.dto.jackson.group.users
 
-fun SuccessfulUsersLookupByEmailResponse.Companion.sample() = SuccessfulUsersLookupByEmailResponse(true, User.sample())
+import io.olaph.slack.dto.jackson.common.types.Member
+import io.olaph.slack.dto.jackson.common.types.sample
+
+fun SuccessfulUsersLookupByEmailResponse.Companion.sample() = SuccessfulUsersLookupByEmailResponse(true, Member.sample())
 
 fun ErrorUsersLookupByEmailResponse.Companion.sample() = ErrorUsersLookupByEmailResponse(false, "")
