+++ data/slack-jackson-dto/src/main/kotlin/io/olaph/slack/dto/jackson/group/users/SetProfile.kt
@@ -57,6 +57,5 @@ data class SetProfileRequest(@JsonProperty("name") val name: String?,
                              @JsonProperty("profile") val profile: Map<String, String>?,
                              @JsonProperty("user") val user: String?,
                              @JsonProperty("value") val value: String?) {
-    companion object {}
-
+    companion object
 }
